Migrate Firebase Analytics to GitLive Firebase-Kotlin-SDK (Phase 7d)
Swaps FirebaseAnalyticsService/AnalyticsModule from the Android-only
com.google.firebase.analytics SDK to dev.gitlive.firebase.analytics, which
is coroutine/multiplatform-first. GitLive's logEvent(name, parameters:
Map<String, Any>?) takes a plain Kotlin map directly, so all the Bundle-
building logic the Android SDK required is removed entirely rather than
ported.

- core/build.gradle.kts: move the Firebase Analytics dependency from
  androidMain (libs.firebase.analytics, Android SDK) to commonMain
  (libs.gitlive.firebase.analytics).
- app/build.gradle.kts: swap libs.firebase.analytics for
  libs.gitlive.firebase.analytics (app has no direct Analytics SDK
  references, only wires the Koin analyticsModule, but needs its own
  dependency since :core's is `implementation`, not `api`).

Also unblocks and moves to commonMain (all were already free of Android
imports once UploadPhoto moved to commonMain in Phase 4b):
- UserDataUseCases.kt and 12 of its 13 constituent use-case files
  (GetAcademicProgress, GetUserData, ObserveUserData, UpdateCreditHours,
  UpdateCumulativeGPA, UpdateDepartment, UpdateFCMToken, UpdateFaculty,
  UpdateLevel, UpdateName, UpdatePhotoUrl, UpdateSemester,
  UpdateUniversity)
- core/domain/user_data/di/Module.kt (the Koin wiring for the above)
- AnalyticsLogger.kt, which depends on UserDataUseCases

SignOut.kt stays in androidMain: it depends on AuthService, whose
signIn(activity: Activity) still leaks android.app.Activity - that
decoupling is deferred to the iOS phase per the migration plan, not
part of this Firebase sub-PR.

Note on verification: unverified in this sandbox - no Android SDK,
Gradle 9.3.1 download blocked by network policy (only 8.14.3 available).
Needs a real ./gradlew build to confirm compilation, plus a real-device/
emulator smoke test that Analytics events still land in Firebase (event
names/params unchanged, only the transport swapped).
