+++ app/build.gradle.kts
@@ -119,13 +119,13 @@ dependencies {
     implementation(libs.androidx.profileinstaller)
     baselineProfile(project(":baselineprofile"))
 
-    // Firebase (Firestore + Storage migrated to GitLive; Auth/Analytics migrate in follow-up
-    // sub-PRs, still on the Android SDK for now)
+    // Firebase (Firestore + Storage + Analytics migrated to GitLive; Auth migrates in a
+    // follow-up sub-PR, still on the Android SDK for now)
     implementation(libs.gitlive.firebase.firestore)
     implementation(libs.gitlive.firebase.storage)
+    implementation(libs.gitlive.firebase.analytics)
     implementation(libs.firebase.auth)
     implementation(libs.firebase.crashlytics)
-    implementation(libs.firebase.analytics)
     implementation(libs.firebase.inappmessaging)
     implementation(libs.firebase.messaging)
     implementation(libs.firebase.perf)
