+++ core/build.gradle.kts
@@ -30,10 +30,13 @@ kotlin {
             // is itself @Composable, which needs the Compose runtime (Composable annotation,
             // stringResource) on the classpath too.
             api(compose.runtime)
+            // GitLive Firebase Analytics (Phase 7d) - replaces the Android-only SDK dependency
+            // that used to live in androidMain, now that FirebaseAnalyticsService/AnalyticsModule
+            // are commonMain.
+            implementation(libs.gitlive.firebase.analytics)
         }
 
         androidMain.dependencies {
-            implementation(libs.firebase.analytics)
             implementation(libs.androidx.core.ktx)
             implementation(libs.androidx.activity.ktx)
             implementation(libs.koin.android)
