+++ core/src/commonMain/kotlin/com/hussienFahmy/core/domain/analytics/AnalyticsLogger.kt
